style: renamed jwt module
#59
